requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-dotenv
gunicorn
psycopg2-binary
whitenoise
